// src/application/domain/weather/weather.enum.ts
export enum Season {
  Spring = "spring",
  Summer = "summer",
  Autumn = "autumn",
  Winter = "winter",
}

export enum TimeOfDay {
  Dawn = "dawn",
  Morning = "morning",
  Noon = "noon",
  Afternoon = "afternoon",
  Sunset = "sunset",
  Dusk = "dusk",
  Night = "night",
  Midnight = "midnight",
}

export enum Condition {
  Sunny = "sunny",
  PartlyCloudy = "partlyCloudy",
  Cloudy = "cloudy",
  Rainy = "rainy",
  Stormy = "stormy",
  Snowy = "snowy",
  Windy = "windy",
  Foggy = "foggy",
  Clear = "clear",
  Humid = "humid",
  Dry = "dry",
  Hot = "hot",
  Cold = "cold",
  Hailing = "hailing",
  Rainbow = "rainbow",
}